Move common table layout logic into `EditorTable` abstract class
