+++ osu.Game/Screens/Edit/EditorTable.cs
@@ -1,10 +1,15 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
+using osu.Framework.Allocation;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
+using osu.Framework.Graphics.Shapes;
+using osu.Framework.Input.Events;
 using osu.Game.Graphics;
+using osu.Game.Graphics.Containers;
 using osu.Game.Graphics.Sprites;
+using osuTK.Graphics;
 
 namespace osu.Game.Screens.Edit
 {
@@ -16,7 +21,7 @@ namespace osu.Game.Screens.Edit
 
         protected const int TEXT_SIZE = 14;
 
-        protected readonly FillFlowContainer BackgroundFlow;
+        protected readonly FillFlowContainer<RowBackground> BackgroundFlow;
 
         protected EditorTable()
         {
@@ -26,7 +31,7 @@ namespace osu.Game.Screens.Edit
             Padding = new MarginPadding { Horizontal = horizontal_inset };
             RowSize = new Dimension(GridSizeMode.Absolute, ROW_HEIGHT);
 
-            AddInternal(BackgroundFlow = new FillFlowContainer
+            AddInternal(BackgroundFlow = new FillFlowContainer<RowBackground>
             {
                 RelativeSizeAxes = Axes.Both,
                 Depth = 1f,
@@ -43,6 +48,92 @@ namespace osu.Game.Screens.Edit
             {
                 Text = text.ToUpper();
                 Font = OsuFont.GetFont(size: 12, weight: FontWeight.Bold);
+            }
+        }
+
+        public class RowBackground : OsuClickableContainer
+        {
+            public readonly object Item;
+
+            private const int fade_duration = 100;
+
+            private readonly Box hoveredBackground;
+
+            [Resolved]
+            private EditorClock clock { get; set; }
+
+            public RowBackground(object item)
+            {
+                Item = item;
+
+                RelativeSizeAxes = Axes.X;
+                Height = 25;
+
+                AlwaysPresent = true;
+
+                CornerRadius = 3;
+                Masking = true;
+
+                Children = new Drawable[]
+                {
+                    hoveredBackground = new Box
+                    {
+                        RelativeSizeAxes = Axes.Both,
+                        Alpha = 0,
+                    },
+                };
+
+                // todo delete
+                Action = () =>
+                {
+                };
+            }
+
+            private Color4 colourHover;
+            private Color4 colourSelected;
+
+            [BackgroundDependencyLoader]
+            private void load(OsuColour colours)
+            {
+                hoveredBackground.Colour = colourHover = colours.BlueDarker;
+                colourSelected = colours.YellowDarker;
+            }
+
+            private bool selected;
+
+            public bool Selected
+            {
+                get => selected;
+                set
+                {
+                    if (value == selected)
+                        return;
+
+                    selected = value;
+                    updateState();
+                }
+            }
+
+            protected override bool OnHover(HoverEvent e)
+            {
+                updateState();
+                return base.OnHover(e);
+            }
+
+            protected override void OnHoverLost(HoverLostEvent e)
+            {
+                updateState();
+                base.OnHoverLost(e);
+            }
+
+            private void updateState()
+            {
+                hoveredBackground.FadeColour(selected ? colourSelected : colourHover, 450, Easing.OutQuint);
+
+                if (selected || IsHovered)
+                    hoveredBackground.FadeIn(fade_duration, Easing.OutQuint);
+                else
+                    hoveredBackground.FadeOut(fade_duration, Easing.OutQuint);
             }
         }
     }
